+++ Audio/ID3v2/V2.cs
@@ -17,7 +17,7 @@ namespace AgEitilt.CardCatalog.Audio.ID3v2 {
 		/// The short name used to represent ID3v2.2 metadata.
 		/// </summary>
 		/// 
-		/// <seealso cref="MetadataFormat.Register{T}(string)"/>
+		/// <seealso cref="FormatRegistry.Register{T}(string)"/>
 		public const string format = "ID3v2.2";
 		/// <summary>
 		/// The display name of the tag format.